postgresSQL/tutorial_container/範例/4台鐵車站進出站範例/requirements.txt
psycopg2-binary>=2.9.3